// src/generated-api/tournament/models/tournament-bracket-match-dto.ts
// tslint:disable
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { TournamentBracketMatchGameDto } from './tournament-bracket-match-game-dto';
import { TournamentParticipantResultDto } from './tournament-participant-result-dto';

/**
 * 
 * @export
 * @interface TournamentBracketMatchDto
 */
export interface TournamentBracketMatchDto {
    /**
     * 
     * @type {number}
     * @memberof TournamentBracketMatchDto
     */
    id: number;
    /**
     * 
     * @type {number}
     * @memberof TournamentBracketMatchDto
     */
    stage_id: number;
    /**
     * 
     * @type {number}
     * @memberof TournamentBracketMatchDto
     */
    group_id: number;
    /**
     * 
     * @type {number}
     * @memberof TournamentBracketMatchDto
     */
    round_id: number;
    /**
     * 
     * @type {number}
     * @memberof TournamentBracketMatchDto
     */
    child_count: number;
    /**
     * 
     * @type {number}
     * @memberof TournamentBracketMatchDto
     */
    number: number;
    /**
     * 
     * @type {string}
     * @memberof TournamentBracketMatchDto
     */
    status: TournamentBracketMatchDtoStatusEnum;
    /**
     * 
     * @type {TournamentParticipantResultDto}
     * @memberof TournamentBracketMatchDto
     */
    opponent1?: TournamentParticipantResultDto;
    /**
     * 
     * @type {TournamentParticipantResultDto}
     * @memberof TournamentBracketMatchDto
     */
    opponent2?: TournamentParticipantResultDto;
    /**
     * 
     * @type {number}
     * @memberof TournamentBracketMatchDto
     */
    startDate: number;
    /**
     * 
     * @type {Array<TournamentBracketMatchGameDto>}
     * @memberof TournamentBracketMatchDto
     */
    games: Array<TournamentBracketMatchGameDto>;
}

/**
    * @export
    * @enum {string}
    */
export enum TournamentBracketMatchDtoStatusEnum {
    Locked = 'Locked',
    Waiting = 'Waiting',
    Ready = 'Ready',
    Running = 'Running',
    Completed = 'Completed',
    Archived = 'Archived'
}