// src/generated-api/tournament/models/tournament-bracket-participant-dto.ts
// tslint:disable
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { TournamentTeamDto } from './tournament-team-dto';

/**
 * 
 * @export
 * @interface TournamentBracketParticipantDto
 */
export interface TournamentBracketParticipantDto {
    /**
     * 
     * @type {number}
     * @memberof TournamentBracketParticipantDto
     */
    id: number;
    /**
     * 
     * @type {number}
     * @memberof TournamentBracketParticipantDto
     */
    tournament_id: number;
    /**
     * 
     * @type {string}
     * @memberof TournamentBracketParticipantDto
     */
    steam_id?: string;
    /**
     * 
     * @type {TournamentTeamDto}
     * @memberof TournamentBracketParticipantDto
     */
    team?: TournamentTeamDto;
}