// src/generated-api/tournament/models/tournament-participant-result-dto.ts
// tslint:disable
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { TournamentBracketParticipantDto } from './tournament-bracket-participant-dto';

/**
 *
 * @export
 * @interface TournamentParticipantResultDto
 */
export interface TournamentParticipantResultDto {
  /**
   *
   * @type {number}
   * @memberof TournamentParticipantResultDto
   */
  id: number | null;
  /**
   *
   * @type {number}
   * @memberof TournamentParticipantResultDto
   */
  position?: number;
  /**
   *
   * @type {boolean}
   * @memberof TournamentParticipantResultDto
   */
  forfeit?: boolean;
  /**
   *
   * @type {number}
   * @memberof TournamentParticipantResultDto
   */
  score?: number;
  /**
   *
   * @type {object}
   * @memberof TournamentParticipantResultDto
   */
  result?: object;
  /**
   *
   * @type {boolean}
   * @memberof TournamentParticipantResultDto
   */
  tbd?: boolean;
  /**
   *
   * @type {TournamentBracketParticipantDto}
   * @memberof TournamentParticipantResultDto
   */
  participant?: TournamentBracketParticipantDto;
}